Multinstance changes commit
